Fix actionlint config YAML parsing in pre-commit
