+++ .github/workflows/security.yml
@@ -64,6 +64,18 @@ jobs:
         id: cargo_audit
         run: cargo audit
 
+
+      - name: Install pinned nightly toolchain for fuzzing
+        run: rustup toolchain install nightly-2025-06-23 --profile minimal
+
+      - name: Install cargo-fuzz
+        run: cargo +nightly-2025-06-23 install cargo-fuzz --locked
+
+      - name: Run fuzz harness smoke
+        id: fuzz_harness
+        working-directory: fuzz
+        run: cargo +nightly-2025-06-23 fuzz run fuzz_passthrough_flags -- -runs=2000
+
       - name: Security summary
         if: always()
         run: |
@@ -72,8 +84,10 @@ jobs:
             echo ""
             echo "- cargo-deny: ${{ steps.cargo_deny.outcome }}"
             echo "- cargo-audit: ${{ steps.cargo_audit.outcome }}"
+            echo "- fuzz harness smoke: ${{ steps.fuzz_harness.outcome }}"
             echo ""
             echo "### Notes"
             echo "- cargo-deny config: \`deny.toml\`"
             echo "- Rust toolchain pinned to 1.88.0 for security checks"
+            echo "- Fuzz toolchain pinned to nightly-2025-06-23 (stable historical nightly)"
           } >> "$GITHUB_STEP_SUMMARY"
