+++ fuzz/fuzz_targets/fuzz_passthrough_flags.rs
@@ -1,0 +1,9 @@
+#![no_main]
+
+use libfuzzer_sys::fuzz_target;
+
+fuzz_target!(|data: &[u8]| {
+    if let Ok(input) = std::str::from_utf8(data) {
+        let _ = windows_mtr::passthrough::parse_passthrough_flags(input);
+    }
+});
